+++ packages/assignment-5/src/origin/__tests__/origin.test.tsx
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
 import { act, fireEvent, render, screen, within } from '@testing-library/react';
-import { CartPage } from '../components/CartPage';
+import { CartPage } from '../components/CartPage.tsx';
 import { Coupon, Product } from '../../types';
 import { AdminPage } from "../components/AdminPage.tsx";
 import { useState } from "react";
