+++ packages/assignment-5/src/refactoring/components/admin/index.ts
@@ -1,0 +1,3 @@
+import { NewProductForm } from "./NewProductForm"
+
+export { NewProductForm }
